+++ tests/src/rules/no-amd.js
@@ -1,11 +1,13 @@
 import { RuleTester } from 'eslint'
+import eslintPkg from 'eslint/package.json'
+import semver from 'semver'
 
 var ruleTester = new RuleTester()
 
 ruleTester.run('no-amd', require('rules/no-amd'), {
   valid: [
-    { code: 'import "x";', parserOptions: { sourceType: 'module' } },
-    { code: 'import x from "x"', parserOptions: { sourceType: 'module' } },
+    { code: 'import "x";', parserOptions: { ecmaVersion: 2015, sourceType: 'module' } },
+    { code: 'import x from "x"', parserOptions: { ecmaVersion: 2015, sourceType: 'module' } },
     'var x = require("x")',
 
     'require("x")',
@@ -25,11 +27,11 @@ ruleTester.run('no-amd', require('rules/no-amd'), {
     'define("a")',
 	],
 
-	invalid: [
-    { code: 'define([], function() {})', errors: [ { message: 'Expected imports instead of AMD define().' }] },
-    { code: 'define(["a"], function(a) { console.log(a); })', errors: [ { message: 'Expected imports instead of AMD define().' }] },
+	invalid: semver.satisfies(eslintPkg.version, '< 4.0.0') ? [] : [
+      { code: 'define([], function() {})', errors: [ { message: 'Expected imports instead of AMD define().' }] },
+      { code: 'define(["a"], function(a) { console.log(a); })', errors: [ { message: 'Expected imports instead of AMD define().' }] },
 
-		{ code: 'require([], function() {})', errors: [ { message: 'Expected imports instead of AMD require().' }] },
-		{ code: 'require(["a"], function(a) { console.log(a); })', errors: [ { message: 'Expected imports instead of AMD require().' }] },
+      { code: 'require([], function() {})', errors: [ { message: 'Expected imports instead of AMD require().' }] },
+      { code: 'require(["a"], function(a) { console.log(a); })', errors: [ { message: 'Expected imports instead of AMD require().' }] },
 	],
 })