+++ tests/src/utils.js
@@ -9,10 +9,26 @@ export function testFilePath(relativePath) {
   return path.join(process.cwd(), './tests/files', relativePath)
 }
 
+export function getTSParsers() {
+  const parsers = []
+  if (semver.satisfies(eslintPkg.version, '>=4.0.0 <6.0.0')) {
+    parsers.push(require.resolve('typescript-eslint-parser'))
+  }
+
+  if (semver.satisfies(eslintPkg.version, '>5.0.0')) {
+    parsers.push(require.resolve('@typescript-eslint/parser'))
+  }
+  return parsers
+}
+
+export function getNonDefaultParsers() {
+  return getTSParsers().concat(require.resolve('babel-eslint'))
+}
+
 export const FILENAME = testFilePath('foo.js')
 
 export function testVersion(specifier, t) {
-  return semver.satisfies(eslintPkg.version, specifier) && test(t)
+  return semver.satisfies(eslintPkg.version, specifier) && test(t())
 }
 
 export function test(t) {
@@ -46,7 +62,7 @@ export const SYNTAX_CASES = [
   test({ code: 'for (let [ foo, bar ] of baz) {}' }),
 
   test({ code: 'const { x, y } = bar' }),
-  test({ code: 'const { x, y, ...z } = bar', parser: 'babel-eslint' }),
+  test({ code: 'const { x, y, ...z } = bar', parser: require.resolve('babel-eslint') }),
 
   // all the exports
   test({ code: 'let x; export { x }' }),
@@ -54,7 +70,7 @@ export const SYNTAX_CASES = [
 
   // not sure about these since they reference a file
   // test({ code: 'export { x } from "./y.js"'}),
-  // test({ code: 'export * as y from "./y.js"', parser: 'babel-eslint'}),
+  // test({ code: 'export * as y from "./y.js"', parser: require.resolve('babel-eslint')}),
 
   test({ code: 'export const x = null' }),
   test({ code: 'export var x = null' }),